+++ CHANGELOG.md
@@ -4,6 +4,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Sign payload for HBS authentication
+
+([#106]): https://github.com/Holo-Host/web-sdk/pull/106
 
 ## [0.6.13-prerelease] - 2023-07-26
 ### Added